implement architecture migration phases 1-3
Phase 1 – Split CometDataInternal.h monolith (1554 lines) into three
focused headers:
  core/Constants.h  – all #define constants and DbType enum
  core/Params.h     – parameter structs (StaticParams, Options, etc.)
  core/Types.h      – runtime types (Query, Results, index structs, externs)
CometDataInternal.h becomes a pure forwarding shim; no .cpp files changed.

Phase 2 – Extract SearchMemoryPool class (threading/SearchMemoryPool.h/.cpp)
Owns the duplicate-fragment scratch arrays previously managed via raw
bool* statics in CometSearch. Uses its own mutex+CV, keeping the FASTA_DB
SearchThreadProc availability array (_pbSearchMemoryPool) as a separate
allocation. FI/PI paths call acquireSlot()/releaseSlot() directly.

Phase 3 – IResultWriter interface + 5 adapter wrappers (output/)
Replaces ~316-line if (bOutputXxx) dispatch chain in DoSearch() with a
factory that builds vector<unique_ptr<IResultWriter>> and calls open(),
write(), close() polymorphically. Each writer owns its file handle(s);
close(bSucceeded, bEmpty) handles format footers, fclose, and remove-on-
empty-search. SQT writer is inserted last to preserve the existing
invariant that WriteSqt runs after all other formats.

All 17 unit tests pass.
