+++ CometSearch/CometDataInternal.h
@@ -12,1543 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// Compatibility shim: existing .cpp files continue to include this header
+// unchanged.  All content has moved to the three focused headers below.
+// New code should include the specific header it needs instead of this one.
 
 #ifndef _COMETDATAINTERNAL_H_
 #define _COMETDATAINTERNAL_H_
 
-#include <chrono>
-#include <string>
-#include <unordered_map>
-#include "CometData.h"
-#include "Threading.h"
-#include "AScoreOptions.h"
-#include "AScoreCentroid.h"
-#include "AScoreAPI.h"
-#include "AScoreFactory.h"
-#include "AScoreDllInterface.h"
-
-
-class CometSearchManager;
-
-#define PROTON_MASS                 1.00727646688
-#define C13_DIFF                    1.00335483
-
-#define FLOAT_ZERO                  1e-6     // 0.000001
-
-#define MIN_PEPTIDE_LEN             1        // min # of AA for a petpide
-#define MAX_PEPTIDE_LEN             51       // max # of AA for a peptide; one more than actual # to account for terminating char
-#define MAX_PEPTIDE_LEN_P2          53       // max # of AA for a peptide plus 2 for N/C-term
-
-#define FRAGINDEX_MIN_IONS_SCORE    3        // min # of matched ions for peptide to register for E-value xcorr histogram
-#define FRAGINDEX_MIN_IONS_REPORT   3        // min # of matched ions for peptide to be reported
-#define FRAGINDEX_MIN_MASS          200.0    // minimum fragment ion mass used to generate fragment index
-#define FRAGINDEX_MAX_MASS          2000.0   // maximum fragment ion mass used to generate fragment index
-#define FRAGINDEX_MAX_BATCHSIZE     1000     // maximum number of spectra loaded when querying fragment index
-#define FRAGINDEX_MAX_NUMPEAKS      150      // number of spectrum peaks used to query fragment index
-#define FRAGINDEX_MAX_NUMSCORED     100      // for each fragment index spectrum query, score up to this many peptides
-#define FRAGINDEX_MAX_COMBINATIONS  2000
-#define FRAGINDEX_MAX_MODS_PER_MOD  5
-#define FRAGINDEX_KEEP_ALL_PEPTIDES 1        // 1 = consider up to FRAGINDEX_MAX_COMBINATIONS of peptides; 0 = ignore all mods for peptide that exceed FRAGINDEX_MAX_COMBINATIONS
-
-#define MS1_MIN_MASS                0.0      // only parse up to this mass in MS1 scans for MS1 library searches
-#define MS1_MAX_MASS                3000.0   // only parse up to this mass in MS1 scans for MS1 library searches
-#define MS1_RT_HISTORY_SIZE         250      // size of MS1 RT history kept for recent history linear regression
-#define MS1_RT_OUTLIER_THRESHOLD    2.0      // # stdev outlier threshold for MS1 RT history
-
-#define MAX_PEFFMOD_LEN             16
-#define SIZE_MASS                   128      // ascii value size
-#define SIZE_NATIVEID               256      // max length of nativeID string
-#define NUM_SP_IONS                 1000     // num ions for preliminary scoring
-#define NUM_ION_SERIES              7        // a,b,c,x,y,z,z1
-#define EXPECT_DECOY_SIZE           3000     // number of decoy entries in CometDecoys.h
-
-#define WIDTH_REFERENCE             256      // length of the protein accession field to store
-#define MAX_PROTEINS                50       // maximum number of proteins to return for each query; for index search only right now
-
-#define HISTO_SIZE                  152      // some number greater than 150
-
-#define NO_PEFF_VARIANT             -127
-
-#define ASCORE_CUTOFF_TO_ACCEPT     13.0     // minimum AScore value to accept localization
-
-#define FRAGINDEX_VMODS             5        // only parse first five variable mods for fragment ion index searches
-                                             // if this is ever larger than 16, need to extend range of siVarModProteinFilter
-
-#define VMODS                       15       // also "VMODS+1" is 4th dimension of uiBinnedIonMasses to cover unmodified ions (0), mod NL (1-15)
-#define COMPOUNDMODS_OFFSET         100      // piVarModSites values >= 100 encode compound mods; index = value - 100
-#define VMOD_1_INDEX                0
-#define VMOD_2_INDEX                1
-#define VMOD_3_INDEX                2
-#define VMOD_4_INDEX                3
-#define VMOD_5_INDEX                4
-#define VMOD_6_INDEX                5
-#define VMOD_7_INDEX                6
-#define VMOD_8_INDEX                7
-#define VMOD_9_INDEX                8
-#define VMOD_10_INDEX               9
-#define VMOD_11_INDEX               10
-#define VMOD_12_INDEX               11
-#define VMOD_13_INDEX               12
-#define VMOD_14_INDEX               13
-#define VMOD_15_INDEX               14
-
-#define ENZYME_SINGLE_TERMINI       1
-#define ENZYME_DOUBLE_TERMINI       2
-#define ENZYME_N_TERMINI            8
-#define ENZYME_C_TERMINI            9
-
-#define ION_SERIES_A                0
-#define ION_SERIES_B                1
-#define ION_SERIES_C                2
-#define ION_SERIES_X                3
-#define ION_SERIES_Y                4
-#define ION_SERIES_Z                5
-#define ION_SERIES_Z1               6  //z+1
-
-#ifdef CRUX
-#define XCORR_CUTOFF                -999.0
-#else
-#define XCORR_CUTOFF                1E-8   // some near-zero cutoff
-#endif
-
-#define SPECLIB_CUTOFF              -999.9
-
-struct Options
-{
-   int iNumPeptideOutputLines;
-   int iWhichReadingFrame;
-   int iEnzymeTermini;
-   int iNumStored;               // # of search results to store for xcorr analysis
-   int iMaxDuplicateProteins;    // maximum number of duplicate proteins to report or store in idx file
-   int iSpectrumBatchSize;       // # of spectra to search at a time within the scan range
-   int iStartCharge;
-   int iEndCharge;
-   int iMaxFragmentCharge;
-   int iMinPrecursorCharge;
-   int iMaxPrecursorCharge;
-   int iMSLevel;                 // filter query scans in raw/mzML/mzXML input by ms level (aka MS2, MS3)
-   int iSpecLibMSLevel;          // filter speclib scans in raw/mzML/mzXML input by ms level (aka MS2, MS3)
-   int iMinPeaks;
-   int iRemovePrecursor;         // 0=no, 1=yes, 2=ETD precursors, 3=phosphate neutral loss
-   int iDecoySearch;             // 0=no, 1=concatenated search, 2=separate decoy search
-   int iNumThreads;              // 0=poll CPU else set # threads to spawn
-   int iNumFragmentThreads;      // # threads used for fragment indexing
-   bool bResolveFullPaths;       // 0=do not resolve full paths; 1=resolve paths (default)
-   bool bOutputSqtStream;
-   bool bOutputSqtFile;
-   bool bOutputTxtFile;
-   bool bOutputPepXMLFile;
-   int iOutputMzIdentMLFile;
-   bool bOutputPercolatorFile;
-   bool bClipNtermMet;           // 0=leave protein sequences alone; 1=also consider w/o N-term methionine
-   bool bClipNtermAA;            // 0=leave peptide sequences as-is; 1=clip N-term amino acid from every peptide
-   bool bMango;                  // 0=normal; 1=Mango x-link ms2 input
-   bool bScaleFragmentNL;        // 0=no; 1=scale fragment NL for each modified residue contained in fragment
-   bool bCreateFragmentIndex;    // 0=normal search; 1=create fragment ion index plain peptide file
-   bool bCreatePeptideIndex;     // 0=normal search; 1=create peptide index file; only one of bCreateFragmentIndex and bCreatePeptideIndex can be 1
-   bool bFastPlainPeptideIdx;    // 0=legacy RunSearch path; 1=use PepGenTuple per-thread buffers (avoids heap alloc)
-   bool bVerboseOutput;
-   bool bExplicitDeltaCn;        // if set to 1, do not use sequence similarity logic
-   bool bPrintExpectScore;
-   bool bExportAdditionalScoresPepXML;  // if 1, also report lnrSp, lnExpect, IonFrac, lnNumSP to pepXML output
-   bool bCorrectMass;            // use selectionMZ instead of monoMZ if monoMZ is outside selection window
-   bool bTreatSameIL;
-   int iPrintAScoreProScore;    // 0=no, otherwise specify variable_modXX number e.g. 1 for variable_mod01
-   int iMaxIndexRunTime;         // max run time of index search in milliseconds
-   int iFragIndexMinIonsScore;   // minimum matched fragment index ions for scoring
-   int iFragIndexMinIonsReport;  // minimum matched fragment index ions for reporting
-   int iFragIndexNumSpectrumPeaks;   // # of peaks from spectrum to use for querying fragment index
-   int iFragIndexSkipReadPrecursors; // if true, skips reading precursors step
-   int iOverrideCharge;
-   long lMaxIterations;          // max # of modification permutations for each iStart position
-   double dMinIntensity;         // intensity cutoff for each peak
-   double dMinPercentageIntensity;   // intensity cutoff for each peak as % of base peak
-   double dRemovePrecursorTol;
-   double dPeptideMassLow;       // MH+ mass
-   double dPeptideMassHigh;      // MH+ mass
-   double dMinimumXcorr;         // set the minimum xcorr to report (default is 1e-8)
-   double dFragIndexMaxMass;     // fragment index maximum fragment mass
-   double dFragIndexMinMass;     // fragment index minimum fragment mass
-   double dMS1MinMass;           // low mass cutoff in MS1 query/library spectra
-   double dMS1MaxMass;           // high mass cutoff in MS1 query/library spectra
-   IntRange scanRange;
-   IntRange peptideLengthRange;
-   DoubleRange clearMzRange;
-   char szActivationMethod[24];  // mzXML only
-   string sPinProteinDelimiter;  // PIN file protein delimiter; default tab
-
-   Options& operator=(Options& a)
-   {
-      iNumPeptideOutputLines = a.iNumPeptideOutputLines;
-      iWhichReadingFrame = a.iWhichReadingFrame;
-      iEnzymeTermini = a.iEnzymeTermini;
-      iNumStored = a.iNumStored;
-      iMaxDuplicateProteins = a.iMaxDuplicateProteins;
-      iSpectrumBatchSize = a.iSpectrumBatchSize;
-      iStartCharge = a.iStartCharge;
-      iEndCharge = a.iEndCharge;
-      iMaxFragmentCharge = a.iMaxFragmentCharge;
-      iMinPrecursorCharge = a.iMinPrecursorCharge;
-      iMaxPrecursorCharge = a.iMaxPrecursorCharge ;
-      iMSLevel = a.iMSLevel;
-      iMinPeaks = a.iMinPeaks;
-      iRemovePrecursor = a.iRemovePrecursor;
-      iDecoySearch = a.iDecoySearch;
-      iNumThreads = a.iNumThreads;
-      bResolveFullPaths = a.bResolveFullPaths;
-      bOutputSqtStream = a.bOutputSqtStream;
-      bOutputSqtFile = a.bOutputSqtFile;
-      bOutputTxtFile = a.bOutputTxtFile;
-      bOutputPepXMLFile = a.bOutputPepXMLFile;
-      iOutputMzIdentMLFile = a.iOutputMzIdentMLFile;
-      bOutputPercolatorFile = a.bOutputPercolatorFile;
-      bClipNtermMet = a.bClipNtermMet;
-      bClipNtermAA = a.bClipNtermAA;
-      bMango = a.bMango;
-      bScaleFragmentNL = a.bScaleFragmentNL;
-      bCreatePeptideIndex = a.bCreatePeptideIndex;
-      bCreateFragmentIndex = a.bCreateFragmentIndex;
-      bFastPlainPeptideIdx = a.bFastPlainPeptideIdx;
-      bVerboseOutput = a.bVerboseOutput;
-      bExplicitDeltaCn = a.bExplicitDeltaCn;
-      bPrintExpectScore = a.bPrintExpectScore;
-      iPrintAScoreProScore = a.iPrintAScoreProScore;
-      bExportAdditionalScoresPepXML = a.bExportAdditionalScoresPepXML;
-      iOverrideCharge = a.iOverrideCharge;
-      bCorrectMass = a.bCorrectMass;
-      bTreatSameIL = a.bTreatSameIL;
-      iMaxIndexRunTime = a.iMaxIndexRunTime;
-      lMaxIterations = a.lMaxIterations;
-      dMinIntensity = a.dMinIntensity;
-      dMinPercentageIntensity = a.dMinPercentageIntensity;
-      dRemovePrecursorTol = a.dRemovePrecursorTol;
-      dPeptideMassLow = a.dPeptideMassLow;
-      dPeptideMassHigh = a.dPeptideMassHigh;
-      dMinimumXcorr = a.dMinimumXcorr;
-      scanRange = a.scanRange;
-      peptideLengthRange = a.peptideLengthRange;
-      clearMzRange = a.clearMzRange;
-      strcpy(szActivationMethod, a.szActivationMethod);
-      sPinProteinDelimiter = a.sPinProteinDelimiter;
-
-      dFragIndexMinMass = a.dFragIndexMinMass;
-      dFragIndexMaxMass = a.dFragIndexMaxMass;
-      iFragIndexMinIonsScore = a.iFragIndexMinIonsScore;    
-      iFragIndexMinIonsReport = a.iFragIndexMinIonsReport ;  
-      iFragIndexNumSpectrumPeaks = a.iFragIndexNumSpectrumPeaks;
-      iFragIndexSkipReadPrecursors = a.iFragIndexSkipReadPrecursors;
-
-      dMS1MinMass = a.dMS1MinMass;
-      dMS1MaxMass = a.dMS1MaxMass;
-
-      return *this;
-   }
-};
-
-struct Results
-{
-   double dPepMass;
-   double dExpect;
-   float  fScoreSp;
-   float  fXcorr;
-   float  fDeltaCn;
-   float  fLastDeltaCn;
-   float  fAScorePro;                         // AScorePro score
-   unsigned short    usiRankXcorr;
-   unsigned short    usiLenPeptide;
-   unsigned short    usiRankSp;
-   unsigned short    usiMatchedIons;
-   unsigned short    usiTotalIons;  
-   comet_fileoffset_t   lProteinFilePosition; // for indexdb, this is the entry in g_pvProteinsList
-   long   lWhichProtein;                      // which entry in g_pvProteinsList[] contains the matched proteins
-   int    piVarModSites[MAX_PEPTIDE_LEN_P2];  // store variable mods encoding, +2 to accomodate N/C-term
-   double pdVarModSites[MAX_PEPTIDE_LEN_P2];  // store variable mods mass diffs, +2 to accomodate N/C-term
-   char   pszMod[MAX_PEPTIDE_LEN][MAX_PEFFMOD_LEN];    // store PEFF mod string
-   char   szPeptide[MAX_PEPTIDE_LEN];
-   char   cPrevAA;                            // stores prev flanking AA
-   char   cNextAA;                            // stores following flanking AA
-   bool   bClippedM;                          // true if new N-term protein due to clipped methionine
-   char   cHasVariableMod;                    // HasVariableModType enum: 0 = no variable mod, 1 = has variable mod, 2 = has AScorePro mod
-   string sPeffOrigResidues;                  // original residue(s) of a PEFF variant
-   string sAScoreProSiteScores;               // AScorePro site scores as comma-separated string
-   int    iPeffOrigResiduePosition;           // position of PEFF variant substitution; -1 = n-term, iLenPeptide = c-term; -9=unused
-   int    iPeffNewResidueCount;               // more than 0 new residues is a substitution (if iPeffOrigResidueCount=1) or insertion (if iPeffOrigResidueCount>1)
-   vector<struct ProteinEntryStruct> pWhichProtein;       // file positions of matched protein entries
-   vector<struct ProteinEntryStruct> pWhichDecoyProtein;  // keep separate decoy list (used for separate decoy matches and combined results)
-};
-
-struct SpecLibResults // MS2 spec lib
-{
-   unsigned int iWhichSpecLib;                // the matched spectral library entry
-   float fSpecLibScore;
-   float fXcorr;                              // use xcorr for now
-   float fCn;                                 // speclib score
-   float fRTtime;                             // retention time in seconds of the matched entry
-};
-
-struct SpecLibResultsMS1  // MS1 spec lib
-{
-   unsigned int iWhichSpecLib;                // the matched spectral library entry
-   float fDotProduct;                         // unit vector dot product aka cosine similarity
-   float fRTime;                              // retention time in seconds of the matched entry
-};
-
-struct PepMassInfo
-{
-   double dCalcPepMass;
-   double dExpPepMass;                        // protonated MH+ experimental mass
-   double dPeptideMassToleranceLow;           // mass tolerance low in amu from experimental mass
-   double dPeptideMassToleranceHigh;          // mass tolerance high in amu from experimental mass
-   double dPeptideMassToleranceMinus;         // low end of mass tolerance range including isotope offsets
-   double dPeptideMassTolerancePlus;          // high end of mass tolerance range including isotope offsets
-};
-
-struct SpectrumInfoInternal
-{
-   int    iArraySize;         // m/z versus intensity array
-   int    iHighestIon;
-   int    iScanNumber;
-   unsigned short    usiChargeState;
-   unsigned short    usiMaxFragCharge;
-   double dTotalIntensity;
-   float  fRTime;
-   char   szMango[32];                // Mango encoding
-   char   szNativeID[SIZE_NATIVEID];  // nativeID string from mzML
-};
-
-// The minimum and maximum mass range of all peptides to consider
-// i.e. lowestPepMass - tolerance to highestPepMass + tolerance
-struct MassRange
-{
-   double dMinMass;
-   double dMaxMass;
-   unsigned short    usiMaxFragmentCharge;  // global maximum fragment charge
-   bool   bNarrowMassRange;    // used to determine how to parse peptides in SearchForPeptides
-   unsigned int uiMaxFragmentArrayIndex; // BIN(dFragIndexMaxMass); used as fragment array index
-};
-
-extern MassRange g_massRange;
-
-// PreprocessStruct stores information used in preprocessing
-// each spectrum.  Information not kept around otherwise
-struct PreprocessStruct
-{
-   int    iHighestIon;
-   double dHighestIntensity;
-};
-
-struct OBOStruct           // stores info read from OBO file
-{
-   double dMassDiffAvg;    // this is looked up from strMod string from OBO
-   double dMassDiffMono;
-   string strMod;          // mod string, PSI-MOD, Unimod or custom
-
-   bool operator<(const OBOStruct& a) const
-   {
-      return (strMod < a.strMod);
-   }
-};
-
-struct ProteinEntryStruct
-{
-   comet_fileoffset_t   lWhichProtein;     // file pointer to protein
-   int    iStartResidue;      // start residue position in protein (1-based)
-   char   cPrevAA;
-   char   cNextAA;
-
-   bool operator<(const ProteinEntryStruct& a) const
-   {
-      return (lWhichProtein < a.lWhichProtein);
-   }
-};
-
-struct PeffModStruct       // stores info read from PEFF header
-{
-   double dMassDiffAvg;    // this is looked up from strMod string from OBO
-   double dMassDiffMono;
-   int    iPosition;       // position of modification
-   char   szMod[MAX_PEFFMOD_LEN];
-
-   bool operator<(const PeffModStruct& a) const
-   {
-      return (iPosition < a.iPosition);
-   }
-};
-
-struct PeffVariantSimpleStruct  // stores info read from PEFF header
-{
-   int    iPosition;       // position of variant
-   char   cResidue;        // new variant
-
-   bool operator<(const PeffVariantSimpleStruct& a) const
-   {
-      return (iPosition < a.iPosition);
-   }
-};
-
-struct PeffVariantComplexStruct  // stores info read from PEFF header
-{
-  int    iPositionA;       // start position of variant
-  int    iPositionB;       // end position of variant
-  string sResidues;        // if !empty(), insertion replacing aa from pos A to B;
-                           // if empty(), deletion of aa from pos A to B
-
-  bool operator<(const PeffVariantComplexStruct& a) const
-  {
-    return (iPositionA < a.iPositionA);
-  }
-};
-
-struct PeffProcessedStruct
-{
-   int iBeginResidue;
-   int iEndResidue;
-};
-
-struct PeffPositionStruct  // collate PEFF mods by position in sequence
-{
-   int iPosition;  // position within the sequence
-   vector<int>    vectorWhichPeff;  // which specific peff entry from PeffModStruct
-   vector<double> vectorMassDiffAvg;
-   vector<double> vectorMassDiffMono;
-};
-
-struct PeffSearchStruct  // variant info passed to SearchForPeptides
-{
-   int    iPosition;
-   bool   bBeginCleavage;
-   bool   bEndCleavage;
-   char   cOrigResidue;
-};
-
-//-->MH
-typedef struct sDBEntry
-{
-   string strName;           // might be able to delete this here
-   string strSeq;
-   comet_fileoffset_t lProteinFilePosition;
-   vector<PeffModStruct> vectorPeffMod;
-   vector<PeffVariantSimpleStruct> vectorPeffVariantSimple;
-   vector<PeffVariantComplexStruct> vectorPeffVariantComplex;
-   vector<PeffProcessedStruct> vectorPeffProcessed;
-} sDBEntry;
-
-struct DBInfo
-{
-   char   szDatabase[SIZE_FILE];
-   char   szFileName[SIZE_FILE];
-   int    iTotalNumProteins;
-   unsigned long int uliTotAACount;
-
-   DBInfo& operator=(DBInfo& a)
-   {
-      strcpy(szDatabase, a.szDatabase);
-      strcpy(szFileName, a.szFileName);
-      iTotalNumProteins = a.iTotalNumProteins;
-      uliTotAACount = a.uliTotAACount;
-
-      return *this;
-   }
-};
-
-struct DBIndex
-{
-   vector<char>          pcVarModSites;                         // empty = unmodified; else [iLen+2] encoding var mods
-   comet_fileoffset_t    lIndexProteinFilePosition;             // points to entry in g_pvProteinsList
-   double                dPepMass;                              // MH+ pep mass
-   unsigned short        siVarModProteinFilter;                 // bitwise representation of mmapProtein
-   char                  cPrevAA;
-   char                  cNextAA;
-   char                  sPeptide[MAX_PEPTIDE_LEN];             // peptide sequence, null-terminated
-
-   bool operator==(const DBIndex& rhs) const
-   {
-      if (strcmp(sPeptide, rhs.sPeptide) != 0)
-         return false;
-
-      if (fabs(dPepMass - rhs.dPepMass) > FLOAT_ZERO)
-         return false;
-
-      int iLen = (int)strlen(sPeptide) + 2;
-      for (int i = 0; i < iLen; ++i)
-      {
-         char l = pcVarModSites.empty()     ? 0 : pcVarModSites[i];
-         char r = rhs.pcVarModSites.empty() ? 0 : rhs.pcVarModSites[i];
-         if (l != r)
-            return false;
-      }
-
-      return true;
-   }
-
-   bool operator<(const DBIndex& rhs) const
-   {
-      int cmp = strcmp(sPeptide, rhs.sPeptide);
-      if (cmp != 0)
-         return cmp < 0;
-
-      if (fabs(dPepMass - rhs.dPepMass) > FLOAT_ZERO)
-         return dPepMass < rhs.dPepMass;
-
-      int iLen = (int)strlen(sPeptide) + 2;
-      for (int i = 0; i < iLen; ++i)
-      {
-         char l = pcVarModSites.empty()     ? 0 : pcVarModSites[i];
-         char r = rhs.pcVarModSites.empty() ? 0 : rhs.pcVarModSites[i];
-         if (l != r)
-            return l < r;
-      }
-
-      // FINAL tie-breaker: lowest protein index first in order
-      // to grab flanking residues from the first protein
-      return lIndexProteinFilePosition < rhs.lIndexProteinFilePosition;
-   }
-};
-
-// Compact fixed-size tuple used during plain-peptide index generation.
-// Replaces heap-heavy DBIndex entries during the per-thread collection phase.
-struct PepGenTuple
-{
-   char     sPeptide[MAX_PEPTIDE_LEN];   // original AA letters (or L->I canonical), null-terminated
-   double   dPepMass;                    // MH+ mass
-   comet_fileoffset_t lProteinFileOffset;// FASTA byte offset of the source protein
-   uint16_t siVarModProteinFilter;
-   char     cPrevAA;
-   char     cNextAA;
-};
-
-// ---------------------------------------------------------------------------
-// 5-bit amino acid encoding for per-length short-peptide key packing.
-// AAs are mapped in ASCII sort order (A=1, C=2, ..., Y=20) so that sorting
-// packed uint64 keys is equivalent to lexicographic sort of sequences within
-// a given peptide length.
-// ---------------------------------------------------------------------------
-static constexpr uint8_t kAA5bit[256] = {
-   0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,   //   0-15
-   0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,   //  16-31
-   0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,   //  32-47
-   0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,   //  48-63
-   0,                                   //  64 '@'
-   1,                                   //  65 'A'
-   0,                                   //  66 'B'
-   2,                                   //  67 'C'
-   3,                                   //  68 'D'
-   4,                                   //  69 'E'
-   5,                                   //  70 'F'
-   6,                                   //  71 'G'
-   7,                                   //  72 'H'
-   8,                                   //  73 'I'  (canonical for I/L when bTreatSameIL)
-   0,                                   //  74 'J'
-   9,                                   //  75 'K'
-  10,                                   //  76 'L'  (remapped to 8 when bTreatSameIL)
-  11,                                   //  77 'M'
-  12,                                   //  78 'N'
-   0,                                   //  79 'O'
-  13,                                   //  80 'P'
-  14,                                   //  81 'Q'
-  15,                                   //  82 'R'
-  16,                                   //  83 'S'
-  17,                                   //  84 'T'
-   0,                                   //  85 'U'
-  18,                                   //  86 'V'
-  19,                                   //  87 'W'
-   0,                                   //  88 'X'
-  20,                                   //  89 'Y'
-   0,                                   //  90 'Z'
-   // 91-255: all zeros
-   0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-   0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-   0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-   0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-   0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-   0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-   0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-   0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-   0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-   0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-   0,0,0,0,0
-};
-
-// Reverse map: 5-bit code -> amino acid character.
-// Code 8 always decodes to 'I' (canonical; L maps to code 8 when bTreatSameIL).
-static constexpr char k5bitAA[32] = {
-   '\0','A','C','D','E','F','G','H','I','K','L','M','N','P','Q','R',
-   'S', 'T','V','W','Y','\0','\0','\0','\0','\0','\0','\0','\0','\0','\0','\0'
-};
-
-// Pack up to 12 amino acids into a uint64 key (5 bits each, 60 bits total).
-// When bTreatSameIL is true, L encodes identically to I.
-inline uint64_t PackPeptide(const char* seq, int iLen, bool bTreatSameIL)
-{
-   uint64_t key = 0;
-   for (int i = 0; i < iLen; ++i)
-   {
-      char c = seq[i];
-      if (bTreatSameIL && c == 'L') c = 'I';
-      key |= ((uint64_t)kAA5bit[(unsigned char)c] << (55 - i * 5));
-   }
-   return key;
-}
-
-// Decode a packed key back to a null-terminated sequence of iLen characters.
-inline void UnpackPeptide(uint64_t key, int iLen, char* seq)
-{
-   for (int i = 0; i < iLen; ++i)
-      seq[i] = k5bitAA[(key >> (55 - i * 5)) & 0x1F];
-   seq[iLen] = '\0';
-}
-
-// Compact per-thread tuple for short peptides (len <= 12) during index generation.
-// 32 bytes on 64-bit (8-byte alignment); uILMask occupies 2 of the 4 trailing pad bytes.
-struct PepGenTupleShort
-{
-   uint64_t           uPackedPep;            // canonical 5-bit-encoded sequence (L treated as I when bTreatSameIL)
-   double             dPepMass;
-   comet_fileoffset_t lProteinFileOffset;
-   uint16_t           siVarModProteinFilter;
-   char               cPrevAA;
-   char               cNextAA;
-   uint16_t           uILMask;              // bitmask: bit k = 1 means position k was 'L' in FASTA original
-};
-
-// This is used for fragment indexing; plain peptides are stored in index
-// file and read in to this data struct.  Same as DBIndex w/o pcVarModSites[]
-struct PlainPeptideIndexStruct
-{
-   comet_fileoffset_t   lIndexProteinFilePosition;  // points to entry in g_pvProteinsList
-   double               dPepMass;                   // MH+ pep mass, unmodified mass; modified mass in FragmentPeptidesStruct
-   unsigned short       siVarModProteinFilter;      // bitwise representation of mmapProtein
-   char                 cPrevAA;
-   char                 cNextAA;
-   char                 szPeptide[MAX_PEPTIDE_LEN]; // peptide sequence, null-terminated
-
-   bool operator==(const PlainPeptideIndexStruct &rhs) const
-   {
-      return strcmp(szPeptide, rhs.szPeptide) == 0;
-   }
-};
-
-struct FragmentPeptidesStruct
-{
-   size_t iWhichPeptide;   // reference to raw peptide (sequence, proteins, etc.) in PlainPeptideIndexStruct
-   int modNumIdx;
-   double dPepMass;     // peptide mass (modified or unmodified) after permuting mods
-   char cNtermMod;
-   char cCtermMod;
-
-   bool operator<(const FragmentPeptidesStruct& a) const
-   {
-      return dPepMass < a.dPepMass;
-   }
-};
-
-struct SpecLibInfo      // why a struct for just a string???
-{
-   string strSpecLibFile;
-};
-
-struct SpecLibStruct
-{
-   string strName;                   // any string associated with speclib entry
-   unsigned int iLibEntry;           // a reference number associated with speclib entry
-   unsigned int iNumPeaks;
-   int iSpecLibCharge;               // precursor charge; not relevant for MS1 speclib
-   double dSpecLibMW;                // if a peptide, store neutral mass
-   float fRTime;
-   float fScaleMinInten;             // min intensity of data prior to encoding to pccSparseFastXcorrData; 0.0 for unit vector
-   float fScaleMaxInten;             // max intensity of data prior to encoding to ppcSparseFastXcorrData
-   vector<std::pair<double, float>> vSpecLibPeaks;
-   float* pfUnitVector;
-   unsigned int uiArraySizeMS1;
-};
-
-// for MS1 alignment
-struct RetentionMatch
-{
-   double dQueryTime;
-   double dReferenceTime;
-   int iSpectrumIndex;
-
-   RetentionMatch(double dQueryTime, double dReferenceTime, int iSpectrumIndex);
-};
-extern std::deque<RetentionMatch> RetentionMatchHistory;
-
-extern unsigned int* g_iFragmentIndex;            // CSR flat data: all posting lists concatenated [g_iFragmentIndexOffset[bin]..g_iFragmentIndexOffset[bin+1])
-extern uint64_t*     g_iFragmentIndexOffset;      // CSR offsets [uiMaxFragmentArrayIndex+1]: cumulative entry counts, can exceed UINT_MAX for large non-enzymatic searches
-extern vector<struct FragmentPeptidesStruct> g_vFragmentPeptides;
-extern vector<PlainPeptideIndexStruct> g_vRawPeptides;
-extern bool* g_bIndexPrecursors;     // allocate an array of BIN(max_precursor, protonated) and use a bool to indicate if that precursor is present in input file(s)
-extern vector<SpecLibStruct> g_vSpecLib;
-extern vector<vector<unsigned int>> g_vulSpecLibPrecursorIndex;  // this will be an vector of vectors<unsigned int>
-
-struct IndexProteinStruct  // for indexed database
-{
-   char szProt[WIDTH_REFERENCE];
-   comet_fileoffset_t lProteinFilePosition;
-   int  iWhichProtein;
-};
-
-struct PEFFInfo
-{
-   char   szPeffOBO[SIZE_FILE];
-   int    iPeffSearch;               // 0=no, 1=PSI-MOD, 2=Unimod, 3=PSI-MOD only, 4=Unimod only, 5=variants only
-};
-
-struct StaticMod
-{
-   double dAddCterminusPeptide;
-   double dAddNterminusPeptide;
-   double dAddCterminusProtein;
-   double dAddNterminusProtein;
-   double pdStaticMods[SIZE_MASS];
-
-   StaticMod& operator=(StaticMod& a)
-   {
-      dAddCterminusPeptide = a.dAddCterminusPeptide;
-      dAddNterminusPeptide = a.dAddNterminusPeptide;
-      dAddCterminusProtein = a.dAddCterminusProtein;
-      dAddNterminusProtein = a.dAddNterminusProtein;
-
-      for (int i = 0; i < SIZE_MASS; ++i)
-      {
-         pdStaticMods[i] = a.pdStaticMods[i];
-      }
-
-      return *this;
-   }
-};
-
-struct PrecalcMasses
-{
-   double dNtermProton;          // dAddNterminusPeptide + PROTON_MASS
-   double dCtermOH2Proton;       // dAddCterminusPeptide + dOH2fragment + PROTON_MASS
-   double dOH2ProtonCtermNterm;  // dOH2parent + PROTON_MASS + dAddCterminusPeptide + dAddNterminusPeptide
-   int    iMinus17;              // BIN'd value of mass(NH3)
-   int    iMinus18;              // BIN'd value of mass(H2O)
-
-   PrecalcMasses& operator=(PrecalcMasses& a)
-   {
-      dNtermProton = a.dNtermProton;
-      dCtermOH2Proton = a.dCtermOH2Proton;
-      dOH2ProtonCtermNterm = a.dOH2ProtonCtermNterm;
-      iMinus17 = a.iMinus17;
-      iMinus18 = a.iMinus18;
-
-      return *this;
-   }
-};
-
-struct VarModParams
-{
-   bool    bVarModSearch;            // set to true if variable mods are specified
-   bool    bVarTermModSearch;        // set to true if any n-term/c-term variable mods are specified
-   bool    bVarProteinNTermMod;      // set to true if a protein n-term variable mod specified
-   bool    bVarProteinCTermMod;      // set to true if a protein c-term variable mod specified 
-   bool    bBinaryModSearch;         // set to true if any of the variable mods are of binary mod variety
-   bool    bUseFragmentNeutralLoss;  // set to true if any custom NL is set; applied only to 1+ and 2+ fragments
-   bool    bRareVarModPresent;       // set to true if any of iRequireThisMod == -1
-   bool    bVarModProteinFilter;     // set to trueif protein mods list is applied
-   int     iRequireVarMod;           // 0=no; else use bits to determine which varmods are required
-   int     iMaxVarModPerPeptide;
-   int     iMaxPermutations;
-   VarMods varModList[VMODS];
-   char    cModCode[VMODS];          // mod characters
-   string  sProteinLModsListFile;                 // file containing list of proteins to restrict application of varmods to
-   multimap<int, string> mmapProteinModsList;     // <varmod#, protein name> vector read from sProteinModsListFile if present
-   string         sCompoundModsFile;              // path to compound mods mass file; empty = disabled
-   vector<double> vdCompoundMasses;               // sorted, deduplicated list of masses read from sCompoundModsFile
-   unsigned int   uiNumCompoundMasses;            // vdCompoundMasses.size(); 0 when feature is disabled
-
-   VarModParams& operator=(VarModParams& a)
-   {
-      bVarModSearch = a.bVarModSearch;
-      bVarTermModSearch = a.bVarTermModSearch;
-      bVarProteinNTermMod = a.bVarProteinNTermMod;
-      bVarProteinCTermMod = a.bVarProteinCTermMod;
-      bBinaryModSearch = a.bBinaryModSearch;
-      bUseFragmentNeutralLoss = a.bUseFragmentNeutralLoss;
-      bRareVarModPresent = a.bRareVarModPresent;
-      bVarModProteinFilter = a.bVarModProteinFilter;
-      iRequireVarMod = a.iRequireVarMod;
-      iMaxVarModPerPeptide = a.iMaxVarModPerPeptide;
-      iMaxPermutations = a.iMaxPermutations;
-
-      for (int i = 0; i < VMODS; ++i)
-      {
-         varModList[i] = a.varModList[i];
-         cModCode[i] = a.cModCode[i];
-      }
-
-      sCompoundModsFile = a.sCompoundModsFile;
-      vdCompoundMasses = a.vdCompoundMasses;
-      uiNumCompoundMasses = a.uiNumCompoundMasses;
-
-      return *this;
-   }
-};
-
-struct MassUtil
-{
-   int    bMonoMassesParent;
-   int    bMonoMassesFragment;
-   double dCO;
-   double dNH3;
-   double dNH2;
-   double dH2O;
-   double dCOminusH2;
-   double dOH2fragment;
-   double dOH2parent;
-   double pdAAMassParent[SIZE_MASS];
-   double pdAAMassFragment[SIZE_MASS];
-   double pdAAMassUser[SIZE_MASS];       // user defined default amino acid masses
-
-   MassUtil& operator=(MassUtil& a)
-   {
-      bMonoMassesParent = a.bMonoMassesParent;
-      bMonoMassesFragment = a.bMonoMassesFragment;
-      dCO = a.dCO;
-      dNH3 = a.dNH3;
-      dNH2 = a.dNH2;
-      dH2O = a.dH2O;
-      dCOminusH2 = a.dCOminusH2;
-      dOH2fragment = a.dOH2fragment;
-      dOH2parent = a.dOH2parent;
-
-      for (int i = 0; i < SIZE_MASS; ++i)
-      {
-         pdAAMassParent[i] = a.pdAAMassParent[i];
-         pdAAMassFragment[i] = a.pdAAMassFragment[i];
-         pdAAMassUser[i] = a.pdAAMassUser[i];
-      }
-
-      return *this;
-   }
-};
-
-struct ToleranceParams
-{
-   int    iMassToleranceUnits;    // 0=amu, 1=mmu, else ppm (2)
-   int    iMassToleranceType;     // 0=MH+ (default), 1=precursor m/z; only valid if iMassToleranceUnits > 0
-   int    iIsotopeError;
-   double dInputToleranceMinus;   // raw tolerance value from param file, lower bound; gets converted to dPeptideMassToleranceMinus
-   double dInputTolerancePlus;    // raw tolerance value from param file, upper bound; gets converted to dPeptideMassTolerancePlus
-   double dFragmentBinSize;
-   double dFragmentBinStartOffset;
-   double dMS1BinSize;
-   double dMS1BinStartOffset;
-
-   ToleranceParams& operator=(ToleranceParams& a)
-   {
-      iMassToleranceUnits = a.iMassToleranceUnits;
-      iMassToleranceType = a.iMassToleranceType;
-      iIsotopeError = a.iIsotopeError;
-      dInputToleranceMinus = a.dInputToleranceMinus;
-      dInputTolerancePlus = a.dInputTolerancePlus;
-      dFragmentBinSize = a.dFragmentBinSize;
-      dFragmentBinStartOffset = a.dFragmentBinStartOffset;
-      dMS1BinSize = a.dMS1BinSize;
-      dMS1BinStartOffset = a.dMS1BinStartOffset;
-
-      return *this;
-   }
-};
-
-struct IonInfo
-{
-   int iNumIonSeriesUsed;
-   int piSelectedIonSeries[NUM_ION_SERIES];
-   bool bUseWaterAmmoniaLoss;    // ammonia, water loss
-   int iTheoreticalFragmentIons;
-   int iIonVal[NUM_ION_SERIES];
-
-   IonInfo& operator=(IonInfo& a)
-   {
-      iNumIonSeriesUsed = a.iNumIonSeriesUsed;
-      bUseWaterAmmoniaLoss = a.bUseWaterAmmoniaLoss;
-      iTheoreticalFragmentIons = a.iTheoreticalFragmentIons;
-
-      for (int i = 0; i < NUM_ION_SERIES; ++i)
-      {
-         piSelectedIonSeries[i] = a.piSelectedIonSeries[i];
-         iIonVal[i] = a.iIonVal[i];
-      }
-
-      return *this;
-   }
-};
-
-// Identifies which type of database is being searched.
-// Defined before StaticParams so iDbType can use DbType.
-enum class DbType
-{
-   FASTA_DB = 0,  // normal FASTA sequence database
-   FI_DB = 1,     // fragment ion index (.idx)
-   PI_DB = 2      // peptide index (.idx)
-};
-
-// static user params, won't change per thread - can make global!
-struct StaticParams
-{
-   string          sHostName;
-   char            szMod[512];         // used for sqt output
-   char            szDecoyPrefix[256]; // used for prefix to indicate decoys
-   string          sDecoyPrefix;       // escaped version of szDecoyPrefix for output within XML files
-   char            szOutputSuffix[256]; // used for suffix to append to output file base names
-   char            szTxtFileExt[256];  // text file extension; default "txt"
-   int             iElapseTime;
-   char            szDate[32];
-   Options         options;
-   DBInfo          databaseInfo;
-   SpecLibInfo     speclibInfo;
-   PEFFInfo        peffInfo;
-   InputFileInfo   inputFile;
-   int             bPrintDuplReferences;
-   VarModParams    variableModParameters;
-   ToleranceParams tolerances;
-   StaticMod       staticModifications;
-   PrecalcMasses   precalcMasses;
-   EnzymeInfo      enzymeInformation;
-   MassUtil        massUtility;
-   double          dInverseBinWidth;    // this is used in BIN() many times so use inverse binWidth to do multiply vs. divide
-   int             iArraySizeGlobal;    // (int)((g_staticParams.options.dPeptideMassHigh + plus_tol_in_daltons + buffer) * g_staticParams.dInverseBinWidth)
-                                        // for MS1 library search, use dMS1MaxMass instead of dPeptideMassHigh
-   double          dOneMinusBinOffset;  // this is used in BIN() many times so calculate once
-   IonInfo         ionInformation;
-   int             iXcorrProcessingOffset;
-   DbType          iDbType;            // FASTA_DB = normal fasta; FI_DB = fragment ion indexed; PI_DB = peptide index
-   vector<double>  vectorMassOffsets;
-   vector<double>  precursorNLIons;
-   int             iPrecursorNLSize;
-   int             iOldModsEncoding;
-   bool            bSkipToStartScan;
-   std::chrono::high_resolution_clock::time_point tRealTimeStart;     // track run time of real-time index search
-
-   StaticParams()
-   {
-       RestoreDefaults();
-   }
-
-   StaticParams& operator=(StaticParams& a)
-   {
-       sHostName = a.sHostName;
-       strcpy(szMod, a.szMod);
-       strcpy(szDecoyPrefix, a.szDecoyPrefix);
-       strcpy(szOutputSuffix, a.szOutputSuffix);
-       strcpy(szTxtFileExt, a.szTxtFileExt);
-       vectorMassOffsets = a.vectorMassOffsets;
-       precursorNLIons= a.precursorNLIons;
-       iPrecursorNLSize = a.iPrecursorNLSize;
-       iOldModsEncoding = a.iOldModsEncoding;
-       iElapseTime = a.iElapseTime;
-       strcpy(szDate, a.szDate);
-       options = a.options;
-       databaseInfo = a.databaseInfo;
-       speclibInfo = a.speclibInfo;
-       inputFile = a.inputFile;
-       bPrintDuplReferences = a.bPrintDuplReferences;
-       variableModParameters = a.variableModParameters;
-       tolerances = a.tolerances;
-       staticModifications = a.staticModifications;
-       precalcMasses = a.precalcMasses;
-       enzymeInformation = a.enzymeInformation;
-       massUtility = a.massUtility;
-       dInverseBinWidth = a.dInverseBinWidth;
-       iArraySizeGlobal = a.iArraySizeGlobal;
-       dOneMinusBinOffset = a.dOneMinusBinOffset;
-       iXcorrProcessingOffset = a.iXcorrProcessingOffset;
-       ionInformation = a.ionInformation;
-       return *this;
-   }
-
-   void RestoreDefaults()
-   {
-      int i;
-
-      inputFile.iInputType = InputType_MS2;
-
-      szMod[0] = '\0';
-
-      iXcorrProcessingOffset = 75;
-      iDbType = DbType::FASTA_DB;
-
-      databaseInfo.szDatabase[0] = '\0';
-      speclibInfo.strSpecLibFile.clear();
-
-      strcpy(szDecoyPrefix, "DECOY_");
-      strcpy(szTxtFileExt, "txt");
-      szOutputSuffix[0] = '\0';
-
-      peffInfo.szPeffOBO[0] = '\0';
-      peffInfo.iPeffSearch = 0;
-
-      variableModParameters.sCompoundModsFile = "";
-      variableModParameters.vdCompoundMasses.clear();
-      variableModParameters.uiNumCompoundMasses = 0;
-
-      iPrecursorNLSize = 0;
-
-      for (i = 0; i < SIZE_MASS; ++i)
-      {
-         massUtility.pdAAMassParent[i] = 999999.;
-         massUtility.pdAAMassFragment[i] = 999999.;
-         massUtility.pdAAMassUser[i] = 0.0;
-         staticModifications.pdStaticMods[i] = 0.0;
-      }
-
-      massUtility.bMonoMassesFragment = 1;
-      massUtility.bMonoMassesParent = 1;
-
-#ifdef CRUX
-      staticModifications.pdStaticMods[(int)'C'] = 57.021464;
-#endif
-
-
-      enzymeInformation.iAllowedMissedCleavage = 2;
-
-      for (i = 0; i < VMODS; ++i)
-      {
-         variableModParameters.varModList[i].iMaxNumVarModAAPerMod = 3;
-         variableModParameters.varModList[i].iMinNumVarModAAPerMod = 0;
-         variableModParameters.varModList[i].iBinaryMod = 0;
-         variableModParameters.varModList[i].iRequireThisMod = 0;
-         variableModParameters.varModList[i].iVarModTermDistance = -1;   // distance from N or C-term distance
-         variableModParameters.varModList[i].iWhichTerm = 0;             // specify N (0) or C-term (1)
-         variableModParameters.varModList[i].dVarModMass = 0.0;
-         variableModParameters.varModList[i].dNeutralLoss = 0.0;
-         variableModParameters.varModList[i].dNeutralLoss2 = 0.0;
-         strcpy(variableModParameters.varModList[i].szVarModChar, "X");
-
-#ifdef CRUX
-         if (i==0)
-         {
-            variableModParameters.varModList[i].dVarModMass = 15.9949;
-            strcpy(variableModParameters.varModList[i].szVarModChar, "M");
-         }
-#endif
-      }
-
-      variableModParameters.cModCode[0] = '*';
-      variableModParameters.cModCode[1] = '#';
-      variableModParameters.cModCode[2] = '@';
-      variableModParameters.cModCode[3] = '^';
-      variableModParameters.cModCode[4] = '~';
-      variableModParameters.cModCode[5] = '$';
-      variableModParameters.cModCode[6] = '%';
-      variableModParameters.cModCode[7] = '!';
-      variableModParameters.cModCode[8] = '+';
-      for (int i = 9; i < VMODS; ++i)
-      {
-         int iAscii = 88 + i;    //start with lower case 'a' ASCII 97
-         if (iAscii <= 125)      // thru '}' which is ASCII 125
-            variableModParameters.cModCode[i] = (char)(iAscii);
-         else
-            variableModParameters.cModCode[i] = '_';
-      }
-
-      variableModParameters.iMaxVarModPerPeptide = 5;
-      variableModParameters.iMaxPermutations = MAX_PERMUTATIONS;
-      variableModParameters.bUseFragmentNeutralLoss = false;
-      variableModParameters.iRequireVarMod = 0;
-
-      ionInformation.bUseWaterAmmoniaLoss = false;
-      ionInformation.iTheoreticalFragmentIons = 1;      // 0 = flanking peaks; 1 = no flanking peaks
-      ionInformation.iIonVal[ION_SERIES_A] = 0;
-      ionInformation.iIonVal[ION_SERIES_B] = 1;
-      ionInformation.iIonVal[ION_SERIES_C] = 0;
-      ionInformation.iIonVal[ION_SERIES_X] = 0;
-      ionInformation.iIonVal[ION_SERIES_Y] = 1;
-      ionInformation.iIonVal[ION_SERIES_Z] = 0;
-      ionInformation.iIonVal[ION_SERIES_Z1] = 0;
-
-      options.iNumPeptideOutputLines = 5;
-      options.iWhichReadingFrame = 0;
-      options.iEnzymeTermini = 2;
-      options.iNumStored = 100;                         // default # of search results to store for xcorr analysis.
-      options.iMaxDuplicateProteins = 20;               // maximum number of duplicate proteins to report or store in idx file
-
-      options.bExplicitDeltaCn = false;
-      options.bPrintExpectScore = true;
-      options.iPrintAScoreProScore = 0;
-      options.bExportAdditionalScoresPepXML = false;
-      options.bCorrectMass = false;
-      options.bTreatSameIL = true;
-      options.iOverrideCharge = 0;
-      options.iMaxIndexRunTime = 0;                     // index run time limit in milliseconds; 0=no time limit
-      options.iRemovePrecursor = 0;
-      options.dRemovePrecursorTol = 1.5;
-
-      options.bOutputSqtStream = false;
-      options.bOutputSqtFile = false;
-      options.bOutputTxtFile = false;
-      options.bOutputPepXMLFile = true;
-      options.iOutputMzIdentMLFile = false;
-      options.bOutputPercolatorFile = false;
-
-      options.bResolveFullPaths = true;
-
-      options.bMango = false;
-      options.bScaleFragmentNL = false;
-      options.bCreatePeptideIndex = false;
-      options.bCreateFragmentIndex = false;
-      options.bFastPlainPeptideIdx = false;
-      options.bVerboseOutput = false;
-      options.iDecoySearch = 0;
-      options.iNumThreads = 4;
-      options.iNumFragmentThreads = 4;
-      options.bClipNtermMet = false;
-      options.bClipNtermAA = false;
-
-      options.lMaxIterations = 0;
-
-      // These parameters affect mzXML/RAMP spectra only.
-      options.scanRange.iStart = 0;
-      options.scanRange.iEnd = 0;
-      options.iSpectrumBatchSize = 0;
-      options.iMinPeaks = 10;
-      options.iStartCharge = 0;
-      options.iEndCharge = 0;
-      options.iMaxFragmentCharge = 3;
-      options.iMinPrecursorCharge = 1;
-      options.iMaxPrecursorCharge = 6;
-      options.iMSLevel = 2;
-      options.dMinIntensity = 0.0;
-      options.dMinPercentageIntensity = 0.0;
-      options.dPeptideMassLow = 600.0;
-      options.dPeptideMassHigh = 5000.0;
-      options.dMinimumXcorr = XCORR_CUTOFF;
-      options.dFragIndexMaxMass = FRAGINDEX_MAX_MASS;
-      options.dFragIndexMinMass = FRAGINDEX_MIN_MASS;
-      strcpy(options.szActivationMethod, "ALL");
-      // End of mzXML specific parameters.
-
-      options.sPinProteinDelimiter = '\t';
-
-      options.dFragIndexMinMass = FRAGINDEX_MIN_MASS;
-      options.dFragIndexMaxMass = FRAGINDEX_MAX_MASS;
-      options.iFragIndexMinIonsScore = FRAGINDEX_MIN_IONS_SCORE;
-      options.iFragIndexMinIonsReport = FRAGINDEX_MIN_IONS_REPORT;
-      options.iFragIndexNumSpectrumPeaks = FRAGINDEX_MAX_NUMPEAKS;
-      options.iFragIndexSkipReadPrecursors = 1;   // skip reading precursors by default
-
-      options.dMS1MinMass = MS1_MIN_MASS;
-      options.dMS1MaxMass = MS1_MAX_MASS;
-
-      options.clearMzRange.dStart = 0.0;
-      options.clearMzRange.dEnd = 0.0;
-
-      options.peptideLengthRange.iStart = MIN_PEPTIDE_LEN;
-      options.peptideLengthRange.iEnd = MAX_PEPTIDE_LEN - 1;  // -1 as MAX_PEPTIDE_LEN number includes terminating char
-
-      staticModifications.dAddCterminusPeptide = 0.0;
-      staticModifications.dAddNterminusPeptide = 0.0;
-      staticModifications.dAddCterminusProtein = 0.0;
-      staticModifications.dAddNterminusProtein = 0.0;
-
-      tolerances.iMassToleranceUnits = 0;
-      tolerances.iMassToleranceType = 0;
-      tolerances.iIsotopeError = 0;
-      tolerances.dInputToleranceMinus = -3.0;               // peptide_mass_tolerance minus
-      tolerances.dInputTolerancePlus = 3.0;                 // peptide_mass_tolerance plus
-      tolerances.dFragmentBinSize = 1.0005;
-      tolerances.dFragmentBinStartOffset = 0.4;
-      tolerances.dMS1BinSize = 1.0005;
-
-      bSkipToStartScan = true;
-   }
-};
-
-extern StaticParams    g_staticParams;
-
-extern vector<DBIndex> g_pvDBIndex;       // used in both peptide index and fragment ion index; latter to store plain peptides
-// Per-length, per-thread generation buffers.  Outer index = (iLen - iMinLen) for short,
-// (iLen - 13) for long.  Inner index = thread slot.
-extern vector<vector<vector<PepGenTupleShort>>> g_vvvPepGenShort;  // lengths <= 12
-extern vector<vector<vector<PepGenTuple>>>      g_vvvPepGenLong;   // lengths > 12
-extern map<long long, IndexProteinStruct>  g_pvProteinNames;   // indexed database protein names and file positions
-
-// Flat CSR-style storage for the per-peptide protein list.
-// Replaces vector<vector<comet_fileoffset_t>> to eliminate the ~190M
-// individual heap allocations (one per inner vector) that caused a
-// ~6-minute free-time tail when building an MHC .idx file.
-// External interface mirrors vector<vector<comet_fileoffset_t>> so
-// existing call sites need no changes.
-class ProteinsListCSR
-{
-public:
-   // Read-only proxy for a single row (one peptide's protein offsets).
-   struct Row
-   {
-      const comet_fileoffset_t* ptr;
-      size_t                    n;
-
-      size_t size()  const { return n; }
-      bool   empty() const { return n == 0; }
-
-      const comet_fileoffset_t& operator[](size_t j) const { return ptr[j]; }
-      comet_fileoffset_t        at(size_t j)          const { return ptr[j]; }
-
-      const comet_fileoffset_t* begin() const { return ptr; }
-      const comet_fileoffset_t* end()   const { return ptr + n; }
-   };
-
-   // Size / state
-   size_t size()  const { return m_off.empty() ? 0 : m_off.size() - 1; }
-   bool   empty() const { return size() == 0; }
-
-   // Modifiers
-   void clear()
-   {
-      vector<comet_fileoffset_t>().swap(m_flat);
-      vector<uint64_t>().swap(m_off);
-   }
-
-   void reserve(size_t n) { m_off.reserve(n + 1); }
-
-   void push_back(const vector<comet_fileoffset_t>& v)
-   {
-      if (m_off.empty()) m_off.push_back(0);
-      m_flat.insert(m_flat.end(), v.begin(), v.end());
-      m_off.push_back(m_flat.size());
-   }
-
-   void push_back(vector<comet_fileoffset_t>&& v)
-   {
-      if (m_off.empty()) m_off.push_back(0);
-      m_flat.insert(m_flat.end(), v.begin(), v.end());
-      m_off.push_back(m_flat.size());
-      vector<comet_fileoffset_t>().swap(v);  // release source buffer immediately
-   }
-
-   // Batch-append from pre-built flat storage.
-   // flat: all protein file offsets for this block, concatenated in row order
-   // cnt:  number of offsets per row (max value bounded by iMaxDuplicateProteins)
-   // Bulk-copies both arrays into m_flat/m_off with two insert() calls, then
-   // releases the source buffers.  Replaces N individual push_back(vector&&)
-   // calls, each of which required one heap free() -- this reduces N free()s
-   // to 2 (one for flat, one for cnt) regardless of how many rows are in the block.
-   void append_flat(vector<comet_fileoffset_t>& flat, vector<uint32_t>& cnt)
-   {
-      if (flat.empty())
-         return;
-      if (m_off.empty())
-         m_off.push_back(0);
-      m_flat.insert(m_flat.end(), flat.begin(), flat.end());
-      for (uint32_t n : cnt)
-         m_off.push_back(m_off.back() + n);
-      vector<comet_fileoffset_t>().swap(flat);
-      vector<uint32_t>().swap(cnt);
-   }
-
-   // Element access
-   Row operator[](size_t i) const
-   {
-      return {m_flat.data() + m_off[i],
-              static_cast<size_t>(m_off[i + 1] - m_off[i])};
-   }
-
-   Row at(size_t i) const { return (*this)[i]; }
-
-   // Range-based for -- yields Row values
-   struct Iterator
-   {
-      const ProteinsListCSR* self;
-      size_t                 i;
-
-      Row       operator*()  const { return (*self)[i]; }
-      Iterator& operator++()       { ++i; return *this; }
-      bool      operator!=(const Iterator& o) const { return i != o.i; }
-   };
-
-   Iterator begin() const { return {this, 0}; }
-   Iterator end()   const { return {this, size()}; }
-
-private:
-   vector<comet_fileoffset_t> m_flat;   // all protein offsets concatenated
-   vector<uint64_t>           m_off;    // [N+1] CSR offsets; row i spans [m_off[i], m_off[i+1])
-};
-
-extern ProteinsListCSR g_pvProteinsList;
-extern unordered_map<comet_fileoffset_t, string> g_pvProteinNameCache;  // file offset -> protein name string; populated at index load
-
-extern std::condition_variable g_searchPoolCV;  // notified when a pool slot is released
-
-extern AScoreProCpp::AScoreOptions g_AScoreOptions;  // AScore options
-extern AScoreProCpp::AScoreDllInterface* g_AScoreInterface;
-
-struct ModificationNumber
-{
-//   int modificationNumber;
-   int modStringLen;             // FIX: need to confirm if not needed  (MOD_SEQS.at(modSeqIdx)).size();
-   char* modifications;
-};
-
-extern vector<ModificationNumber> MOD_NUMBERS;
-extern vector<string> MOD_SEQS;    // Unique modifiable sequences.
-extern int* MOD_SEQ_MOD_NUM_START; // Start index in the MOD_NUMBERS vector for a modifiable sequence; -1 if no modification numbers were generated
-extern int* MOD_SEQ_MOD_NUM_CNT;   // Total modifications numbers for a modifiable sequence.
-
-// Index into the MOD_SEQS vector
-// -1 for peptides that have no modifiable amino acids
-// -2 for peptides with no modifiable amino acids but contain n/c-term mods
-extern int* PEPTIDE_MOD_SEQ_IDXS;
-
-extern int MOD_NUM;
-extern bool g_bPlainPeptideIndexRead;   // set to true if plain peptide index file is read (and fragment index generated)
-                                        // poor choice of name for the fragment index .idx given peptide index is back
-extern  std::atomic<bool>  g_bPeptideIndexRead;        // set to true if peptide index file is read
-extern bool g_bSpecLibRead;             // set to true if spectral library file is read
-
-extern bool g_bPerformSpecLibSearch;    // set to true if doing spectral library search
-extern bool g_bPerformDatabaseSearch;   // set to true if doing database search
-
-extern bool g_bCometPreprocessMemoryAllocated;    // set to true when memory has been allocated
-extern bool g_bCometSearchMemoryAllocated;        // set to true when memory has been allocated
-
-extern bool g_bIdxNoFasta;   // set to true when .idx file being search but corresponding .fasta not present
-                             // used in mzid output to skip sequence retrieval
-
-// Query stores information for peptide scoring and results
-// This struct is allocated for each spectrum/charge combination
-struct Query
-{
-   int   iXcorrHistogram[HISTO_SIZE];
-   unsigned int   uiHistogramCount;   // # of entries in histogram
-   float fPar[4];           // parameters of LMA regression
-
-   int iMatchPeptideCount;        // # of peptides that get stored (i.e. are greater than lowest score)
-   int iDecoyMatchPeptideCount;   // # of decoy peptides that get stored (i.e. are greater than lowest score)
-
-   short siMaxXcorr;        // index of maximum correlation score in iXcorrHistogram
-
-   short siLowestXcorrScoreIndex;
-   short siLowestDecoyXcorrScoreIndex;
-
-   double dLowestXcorrScore;
-   double dLowestDecoyXcorrScore;
-
-   float fLowestSpecLibScore;
-
-   int iMinXcorrHisto;    // min xcorr score for xcorr histogram to address good E-values for poor/sparse spectra
-
-   double dMangoIndex;      // scan number decimal precursor value i.e. 2401.001 for scan 2401, first precursor/z pair
-
-   unsigned long int  _uliNumMatchedPeptides;  // # of peptides that get scored
-   unsigned long int  _uliNumMatchedDecoyPeptides;
-
-   // When true, sparse child arrays (float[SPARSE_MATRIX_SIZE]) belong to the
-   // thread-local RtsScratch pool and must NOT be delete[]'d by the destructor.
-   // Set only by PreprocessSingleSpectrumThreadLocal via PreprocessSingleSpectrumCore.
-   bool bSparseFromPool;
-
-   // Sparse matrix representation of data
-   int iSpScoreData;    //size of sparse matrix
-   int iFastXcorrDataSize;
-   float **ppfSparseSpScoreData;
-   float **ppfSparseFastXcorrData;
-   float **ppfSparseFastXcorrDataNL;  // ppfSparseFastXcorrData with NH3, H2O contributions
-
-   // Store raw peaks for AScorePro
-
-   // List of ms/ms masses for fragment index search; intensity not important at this stage
-   vector<float> vfRawFragmentPeakMass;
-   // Consider replacing vfRawFragmentPeakMass with a vector<pair<double, double>> to store
-   // both mass and intensity if AScorePro is used
-   vector<AScoreProCpp::Centroid> vRawFragmentPeakMassIntensity;
-
-
-   PepMassInfo          _pepMassInfo;
-   SpectrumInfoInternal _spectrumInfoInternal;
-   Results*             _pResults;
-   Results*             _pDecoys;
-   SpecLibResults*      _pSpecLibResults;
-
-   std::chrono::high_resolution_clock::time_point tSearchStart;  // per-query search start time for iMaxIndexRunTime timeout
-
-   Mutex accessMutex;
-
-   Query()
-   {
-      memset(iXcorrHistogram, 0, sizeof(iXcorrHistogram));
-
-      iMatchPeptideCount = 0;
-      iDecoyMatchPeptideCount = 0;
-      uiHistogramCount = 0;
-      iMinXcorrHisto = 0;
-
-      fPar[0]=0.0;
-      fPar[1]=0.0;
-      fPar[2]=0.0;
-      fPar[3]=0.0;
-
-      siMaxXcorr = 0;                        // index of maximum correlation score in iXcorrHistogram
-      siLowestXcorrScoreIndex = 0;
-      siLowestDecoyXcorrScoreIndex = 0;
-
-      dLowestXcorrScore = XCORR_CUTOFF;
-      dLowestDecoyXcorrScore = XCORR_CUTOFF;
-
-      fLowestSpecLibScore = SPECLIB_CUTOFF;
-
-      dMangoIndex = 0.0;
-
-      _uliNumMatchedPeptides = 0;
-      _uliNumMatchedDecoyPeptides = 0;
-
-      bSparseFromPool = false;
-
-      ppfSparseSpScoreData = NULL;
-      ppfSparseFastXcorrData = NULL;
-      ppfSparseFastXcorrDataNL = NULL;          // ppfSparseFastXcorrData with NH3, H2O contributions
-
-      vfRawFragmentPeakMass.clear();
-      vRawFragmentPeakMassIntensity.clear();
-
-      _pepMassInfo.dCalcPepMass = 0.0;
-      _pepMassInfo.dExpPepMass = 0.0;
-      _pepMassInfo.dPeptideMassToleranceLow = 0.0;
-      _pepMassInfo.dPeptideMassToleranceHigh = 0.0;
-      _pepMassInfo.dPeptideMassToleranceMinus = 0.0;
-      _pepMassInfo.dPeptideMassTolerancePlus = 0.0;
-
-      _spectrumInfoInternal.dTotalIntensity = 0.0;
-      _spectrumInfoInternal.iArraySize = 0;
-      _spectrumInfoInternal.iHighestIon = 0;
-      _spectrumInfoInternal.iScanNumber = 0;
-      _spectrumInfoInternal.dTotalIntensity = 0.0;
-
-      _pResults = NULL;
-      _pDecoys = NULL;
-      _pSpecLibResults = NULL;
-
-      Threading::InitMutex(&accessMutex);
-   }
-
-   ~Query()
-   {
-      int i;
-      if (!bSparseFromPool)
-      {
-         for (i = 0; i < iSpScoreData; ++i)
-         {
-            if (ppfSparseSpScoreData[i] != NULL)
-               delete[] ppfSparseSpScoreData[i];
-         }
-      }
-      delete[] ppfSparseSpScoreData;
-      ppfSparseSpScoreData = NULL;
-
-      if (g_staticParams.ionInformation.bUseWaterAmmoniaLoss
-            && (g_staticParams.ionInformation.iIonVal[ION_SERIES_A]
-               || g_staticParams.ionInformation.iIonVal[ION_SERIES_B]
-               || g_staticParams.ionInformation.iIonVal[ION_SERIES_Y]))
-      {
-         if (!bSparseFromPool)
-         {
-            for (i = 0; i < iFastXcorrDataSize; ++i)
-            {
-               if (ppfSparseFastXcorrData[i] != NULL)
-                  delete[] ppfSparseFastXcorrData[i];
-               if (ppfSparseFastXcorrDataNL[i]!=NULL)
-                  delete[] ppfSparseFastXcorrDataNL[i];
-            }
-         }
-         delete[] ppfSparseFastXcorrDataNL;
-         ppfSparseFastXcorrDataNL = NULL;
-      }
-      else
-      {
-         if (!bSparseFromPool)
-         {
-            for (i = 0; i < iFastXcorrDataSize; ++i)
-            {
-               if (ppfSparseFastXcorrData[i] != NULL)
-                  delete[] ppfSparseFastXcorrData[i];
-            }
-         }
-      }
-      delete[] ppfSparseFastXcorrData;
-      ppfSparseFastXcorrData = NULL;
-
-      if (_pResults != NULL)
-      {
-         _pResults->pWhichProtein.clear();
-         if (g_staticParams.options.iDecoySearch == 1)
-            _pResults->pWhichDecoyProtein.clear();
-         delete[] _pResults;
-         _pResults = NULL;
-      }
-
-      if (g_staticParams.options.iDecoySearch == 2 && _pDecoys != NULL)
-      {
-         _pDecoys->pWhichDecoyProtein.clear();
-         delete[] _pDecoys;
-         _pDecoys = NULL;
-      }
-
-      Threading::DestroyMutex(accessMutex);
-   }
-};
-
-struct QueryMS1
-{
-   //   short siLowestSpecLibIndex;
-   //   float fLowestXcorr;
-   unsigned int uiMatchMS1Count;        // # of peptides that get stored (i.e. are greater than lowest score)
-   unsigned int iArraySizeMS1;          // dimension of pcFastXcorrData
-
-   // Standard array representation of data
-   // Library spectra are fast xcorr manipulated so non need to do so with query MS1
-   float* pfFastXcorrData;
-
-   SpecLibResultsMS1 _pSpecLibResultsMS1;
-
-   Mutex accessMutex;
-
-   QueryMS1()
-   {
-      //      siLowestSpecLibIndex = 0;
-      //      fLowestXcorr = SPECLIB_CUTOFF;
-      uiMatchMS1Count = 0;
-      pfFastXcorrData = NULL;
-      _pSpecLibResultsMS1.fDotProduct = 0.0;
-      _pSpecLibResultsMS1.fRTime = 0.0;
-
-      Threading::InitMutex(&accessMutex);
-   }
-
-   ~QueryMS1()
-   {
-      //FIX delete _pSepcLibResults
-
-      Threading::DestroyMutex(accessMutex);
-   }
-};
-
-extern vector<Query*>          g_pvQuery;
-extern vector<QueryMS1*>       g_pvQueryMS1;
-extern vector<InputFileInfo*>  g_pvInputFiles;
-extern Mutex                   g_pvQueryMutex;
-extern Mutex                   g_pvDBIndexMutex;
-extern Mutex                   g_preprocessMemoryPoolMutex;
-extern Mutex                   g_searchMemoryPoolMutex;
-extern Mutex                   g_dbIndexMutex;
-extern Mutex                   g_vSpecLibMutex;
-
-struct IonSeriesStruct         // defines which fragment ion series are considered
-{
-   int bPreviousMatch[8];
-};
-
-
-struct MatchedIonsStruct  // for SingleSpectrumSearch
-{
-   double dMass;
-   double dInten;
-
-   bool operator<(const MatchedIonsStruct& a) const
-   {
-      return dInten > a.dInten;
-   }
-};
+#include "core/Constants.h"
+#include "core/Params.h"
+#include "core/Types.h"
 
 #endif // _COMETDATAINTERNAL_H_
